Refactor setUp() methodes in test classes
