+++ tests/test_disaggregation.py
@@ -2,26 +2,13 @@ import ast
 import json
 from unittest import mock
 from flask_api import status
-from models.user import User, GenderType, StateType
-from models.group import Group
 from tests.buzzn_test_case import BuzznTestCase
 from tests.string_constants import EMPTY_RESPONSE, DISAGGREGATION,\
     INDIVIDUAL_DISAGGREGATION
-from util.database import db
 
 
 class IndividualDisaggregation(BuzznTestCase):
     """ Unit tests for IndividualDisaggregation. """
-
-    def setUp(self):
-        super().setUp()
-        self.target_user = User(GenderType.MALE, "Some", "User", "user@some.net",
-                                "SomeToken", "SomeMeterId", "SomeGroup")
-        self.target_user.set_password("some_password")
-        self.target_user.state = StateType.ACTIVE
-        self.target_user.meter_id = 'EASYMETER_1124001747'
-        db.session.add(self.target_user)
-        db.session.commit()
 
     # pylint does not understand the required argument from the @mock.patch decorator
     # pylint: disable=unused-argument
@@ -31,7 +18,7 @@ class IndividualDisaggregation(BuzznTestCase):
 
         # Check if route exists
         login_request = self.client.post('/login',
-                                         data=json.dumps({'user': 'User@Some.net',
+                                         data=json.dumps({'user': 'test@test.net',
                                                           'password': 'some_password'}))
         response = self.client.get('/individual-disaggregation', headers={
             'Authorization': 'Bearer {}'.format(login_request.json["sessionToken"])})
@@ -48,23 +35,6 @@ class IndividualDisaggregation(BuzznTestCase):
 class GroupDisaggregation(BuzznTestCase):
     """ Unit tests for GroupDisaggregation. """
 
-    def setUp(self):
-        super().setUp()
-        self.target_user = User(GenderType.MALE, "Some", "User", "user@some.net",
-                                "SomeToken", "SomeMeterId", "SomeGroup")
-        self.target_user.set_password("some_password")
-        self.target_user.state = StateType.ACTIVE
-        self.target_user.meter_id = 'EASYMETER_60404854'
-        self.target_user.group_id = 1
-        db.session.add(self.target_user)
-        self.target_group = Group(
-            "SomeGroup",
-            'EASYMETER_1124001747',
-            '5e769d5b83934bccae11a8fa95e0dc5f',
-            'e2a7468f0cf64b7ca3f3d1350b893c6d')
-        db.session.add(self.target_group)
-        db.session.commit()
-
     # pylint does not understand the required argument from the @mock.patch decorator
     # pylint: disable=unused-argument
     @mock.patch('routes.disaggregation.get_default_disaggregation', return_value=EMPTY_RESPONSE)
@@ -73,7 +43,7 @@ class GroupDisaggregation(BuzznTestCase):
 
         # Check if route exists
         login_request = self.client.post('/login',
-                                         data=json.dumps({'user': 'User@Some.net',
+                                         data=json.dumps({'user': 'test@test.net',
                                                           'password': 'some_password'}))
         response = self.client.get('/group-disaggregation', headers={
             'Authorization': 'Bearer {}'.format(login_request.json["sessionToken"])})
@@ -91,23 +61,6 @@ class Disaggregation(BuzznTestCase):
     """ Unit test for common funcionalities of IndividualDisaggregation and
     GroupDisaggregation. """
 
-    def setUp(self):
-        super().setUp()
-        self.target_user = User(GenderType.MALE, "Some", "User", "user@some.net",
-                                "SomeToken", "SomeMeterId", "SomeGroup")
-        self.target_user.set_password("some_password")
-        self.target_user.state = StateType.ACTIVE
-        self.target_user.meter_id = 'EASYMETER_60404854'
-        self.target_user.group_id = 1
-        db.session.add(self.target_user)
-        self.target_group = Group(
-            'SomeGroup',
-            '52d7c87f8c26433dbd095048ad30c8cf',
-            'EASYMETER_60404852',
-            'EASYMETER_60327599')
-        db.session.add(self.target_group)
-        db.session.commit()
-
     # pylint: disable=unused-argument
     @mock.patch('routes.disaggregation.get_disaggregation',
                 return_value=EMPTY_RESPONSE)
@@ -117,7 +70,7 @@ class Disaggregation(BuzznTestCase):
         """ Test handling of erroneous parameters. """
 
         login_request = self.client.post('/login',
-                                         data=json.dumps({'user': 'User@Some.net',
+                                         data=json.dumps({'user': 'test@test.net',
                                                           'password': 'some_password'}))
 
         for route in '/individual-disaggregation', '/group-disaggregation':
