+++ tests/test_group_profile_pictures.py
@@ -16,15 +16,6 @@ class GroupProfilePictures(BuzznTestCase):
 
     def setUp(self):
         super().setUp()
-        db.session.add(Group('TestGroup',
-                             '269e682dbfd74a569ff4561b6416c999',
-                             '5e769d5b83934bccae11a8fa95e0dc5f',
-                             'e2a7468f0cf64b7ca3f3d1350b893c6d'))
-        test_user = User(GenderType.MALE, 'Some', 'User', 'test@test.net', 'TestToken',
-                         'b4234cd4bed143a6b9bd09e347e17d34', 1)
-        test_user.set_password('some_password')
-        test_user.state = StateType.ACTIVE
-        db.session.add(test_user)
         group_member2 = User(GenderType.FEMALE, 'judith', 'greif',
                              'judith@buzzn.net', 'TestToken2',
                              '52d7c87f8c26433dbd095048ad30c8cf', 1)
@@ -35,7 +26,6 @@ class GroupProfilePictures(BuzznTestCase):
                              '117154df05874f41bfdaebcae6abfe98', 1)
         group_member3.set_password('some_password3')
         group_member3.state = StateType.ACTIVE
-        db.session.add(test_user)
         db.session.add(group_member2)
         db.session.add(group_member3)
         db.session.commit()
